# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Sales Predictor API
  version: 1.0.0
  description: API for sales forecasting, inventory management, and market basket analysis
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
          description: Report ID
          example: "507f191e810c19729de860ee"
        userId:
          type: string
          description: User who generated the report
          example: "507f1f77bcf86cd799439011"
        reportType:
          type: string
          enum: [Sales, Forecast, Inventory, Performance]
          description: Type of report
          example: "Sales"
        format:
          type: string
          enum: [csv, pdf]
          description: File format
          example: "pdf"
        filePath:
          type: string
          description: Path to report file
          example: "/reports/Sales_1623456789012.pdf"
        dateRange:
          type: object
          properties:
            startDate:
              type: string
              format: date-time
              description: Start date of data range
              example: "2025-05-01T00:00:00Z"
            endDate:
              type: string
              format: date-time
              description: End date of data range
              example: "2025-05-31T23:59:59Z"
          example:
            startDate: "2025-05-01T00:00:00Z"
            endDate: "2025-05-31T23:59:59Z"
      required: [userId, reportType, format, filePath, dateRange]
...
paths:
...
  /api/reports/generate:
    post:
      summary: Generate report
      tags: [Reports]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reportType:
                  type: string
                  enum: [Sales, Forecast, Inventory, Performance]
                  example: "Sales"
                format:
                  type: string
                  enum: [csv, pdf]
                  example: "pdf"
                startDate:
                  type: string
                  format: date-time
                  example: "2025-05-01T00:00:00Z"
                endDate:
                  type: string
                  format: date-time
                  example: "2025-05-31T23:59:59Z"
              required: [reportType, format, startDate, endDate]
      responses:
        "201":
          description: Report generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report:
                    $ref: "#/components/schemas/Report"
                  downloadLink:
                    type: string
                    example: "/reports/Sales_1623456789012.pdf"
              example:
                report:
                  id: "507f191e810c19729de860ee"
                  userId: "507f1f77bcf86cd799439011"
                  reportType: "Sales"
                  format: "pdf"
                  filePath: "/reports/Sales_1623456789012.pdf"
                  dateRange:
                    startDate: "2025-05-01T00:00:00Z"
                    endDate: "2025-05-31T23:59:59Z"
                downloadLink: "/reports/Sales_1623456789012.pdf"
        "400":
          description: No data or validation error
          content:
            application/json:
              example:
                error: "No data for the selected range"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
        "403":
          description: Access denied (non-Manager, non-Planner, non-Owner, non-Admin)
          content:
            application/json:
              example:
                error: "Access denied"
  /api/reports:
    get:
      summary: List reports
      tags: [Reports]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: reportType
          schema:
            type: string
            enum: [Sales, Forecast, Inventory, Performance]
            example: "Sales"
          description: Filter by report type
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
              example:
                - id: "507f191e810c19729de860ee"
                  userId: "507f1f77bcf86cd799439011"
                  reportType: "Sales"
                  format: "pdf"
                  filePath: "/reports/Sales_1623456789012.pdf"
                  dateRange:
                    startDate: "2025-05-01T00:00:00Z"
                    endDate: "2025-05-31T23:59:59Z"
                - id: "507f191e810c19729de860ef"
                  userId: "507f1f77bcf86cd799439011"
                  reportType: "Inventory"
                  format: "csv"
                  filePath: "/reports/Inventory_1623456789013.csv"
                  dateRange:
                    startDate: "2025-05-01T00:00:00Z"
                    endDate: "2025-05-31T23:59:59Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              example:
                error: "Unauthorized"
...